test: split LLMFS generation suite
